SeleniumFramework/requirements.txt
allure-pytest == 2.7.0
allure-python-commons == 2.7.0
attrs  == 23.2.0
certifi == 2024.2.2
chardet == 3.0.4
colorama  ==  0.4.6
et-xmlfile  ==  1.1.0
exceptiongroup == 1.2.0
idna ==  2.8
importlib-metadata ==  6.7.0
iniconfig  ==  2.0.0
jdcal  ==  1.4.1
MouseInfo  == 0.1.3
openpyxl  == 2.6.2
packaging == 23.2
Pillow == 9.5.0
pip == 24.0
pluggy == 1.2.0
PyAutoGUI == 0.9.54
PyGetWindow == 0.0.9
pyjnius == 1.6.1
PyMsgBox == 1.0.9
pyperclip == 1.8.2
PyRect == 0.2.0
PyScreeze == 0.1.30
pytest ==  7.4.4
python-dotenv == 0.21.1
pytweening == 1.2.0
requests == 2.22.0
selenium == 3.141.0
setuptools == 40.8.0
six == 1.12.0
tomli == 2.0.1
typing_extensions == 4.7.1
urllib3 == 1.25.3
webdriver-manager  == 4.0.1
zipp == 3.15.0

